backend/requirements.txt
fastapi
uvicorn
motor
beanie
dnspython
pydantic
pydantic-settings
python-dotenv
httpx
jinja2
email-validator
python-multipart
openai>=1.0.0
tiktoken
tenacity
pypdf
PyPDF2
python-docx
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
requests
